requirements.txt
streamlit
streamlit[pdf]
pandas
requests 
pymupdf 
openpyxl
selenium
sqlalchemy
psycopg2-binary
python-dotenv
pgvector
sentence-transformers
torchvision